+++ example/lib/widgets/setting/options.dart
@@ -4,7 +4,7 @@ import 'package:flutter_event_calendar/flutter_event_calendar.dart';
 import 'package:flutter_event_calendar_example/injection.dart';
 import 'package:flutter_event_calendar_example/models/calendar_event_model.dart';
 import 'package:flutter_event_calendar_example/models/expanded_item_model.dart';
-import 'package:flutter_event_calendar_example/widgets/setting/data_picker.dart';
+import 'package:flutter_event_calendar_example/widgets/setting/date_picker.dart';
 import 'package:flutter_event_calendar_example/widgets/setting/item_setting.dart';
 
 import 'package:flutter_event_calendar_example/widgets/color/calendar_color_item.dart';
@@ -42,7 +42,7 @@ class _OptionsState extends State<Options> {
       ExpandedItemModel(
           icon: 'assets/language_icon.svg',
           title: "Special Days",
-          body: SelectDatePicker(
+          body: DatePicker(
             onDayChanged: () {
              widget.onDayChanged?.call(EventCalendar.dateTime!);
              setState(() {});
